Refactor the "Data was entered by humans" section
This paragraph is now shorter and less curcuitous.

Original version claimed "There is no worse way to screw up data than to let a single human type it in".  This is false.  Data variability increases significantly with the number of people entering data in free form fields.

Also the claim that "Even with the best tools available, data this messy can't be saved" is overblown.  The example, as described, isn't difficult since the data domain (dog breeds) is finite.

It might take time, but finite domain data is not intractable data cleansing problem.  Therefore no need for drama abd extreme conclusions from this particular example.